Remove BaseActivity
