+++ app/src/main/java/org/frknkrc44/hma_oss/ui/activity/MainActivity.kt
@@ -1,19 +1,45 @@
 package org.frknkrc44.hma_oss.ui.activity
 
+import android.content.Context
 import android.content.res.Configuration
+import android.content.res.Resources
 import android.os.Bundle
+import androidx.activity.enableEdgeToEdge
+import androidx.appcompat.app.AppCompatActivity
+import androidx.core.graphics.drawable.toDrawable
 import androidx.navigation.findNavController
+import com.google.android.material.color.DynamicColors
+import com.google.android.material.color.DynamicColorsOptions
+import icu.nullptr.hidemyapplist.hmaApp
+import icu.nullptr.hidemyapplist.service.PrefManager
 import icu.nullptr.hidemyapplist.service.ServiceClient
+import icu.nullptr.hidemyapplist.ui.util.ThemeUtils
+import icu.nullptr.hidemyapplist.util.ConfigUtils
 import org.frknkrc44.hma_oss.BuildConfig
 import org.frknkrc44.hma_oss.R
 import org.frknkrc44.hma_oss.databinding.ActivityMainBinding
 
-class MainActivity : BaseActivity() {
+class MainActivity : AppCompatActivity() {
     var readyToKill: Boolean = true
     var currentConfiguration: Configuration? = null
 
     override fun onCreate(savedInstanceState: Bundle?) {
         super.onCreate(savedInstanceState)
+
+        enableEdgeToEdge()
+
+        // I add this manually because the E2E code is not working like I want
+        // They should give us a separate method to set it for enableEdgeToEdge
+        // Source: https://github.com/androidx/androidx/blob/c0f9aabcf6f32029249ac7647711744b68e2a003/activity/activity/src/main/java/androidx/activity/EdgeToEdge.kt#L299
+        window.isNavigationBarContrastEnforced = !PrefManager.systemWallpaper
+
+        DynamicColors.applyToActivityIfAvailable(
+            this,
+            DynamicColorsOptions.Builder().also {
+                if (!ThemeUtils.isSystemAccent)
+                    it.setThemeOverlay(ThemeUtils.getColorThemeStyleRes(this))
+            }.build()
+        )
 
         currentConfiguration = resources.configuration
 
@@ -32,11 +58,24 @@ class MainActivity : BaseActivity() {
         currentConfiguration = newConfig
     }
 
-    private fun isNightModeEnabled(config: Configuration?) = config?.uiMode?.and(Configuration.UI_MODE_NIGHT_MASK)
+    override fun onApplyThemeResource(theme: Resources.Theme, resid: Int, first: Boolean) {
+        super.onApplyThemeResource(theme, resid, first)
+        if (!DynamicColors.isDynamicColorAvailable()) {
+            theme.applyStyle(ThemeUtils.getColorThemeStyleRes(this), true)
+        }
+
+        theme.applyStyle(ThemeUtils.getOverlayThemeStyleRes(this), true)
+
+        applyWallpaperBackgroundColor()
+    }
 
     override fun onSupportNavigateUp(): Boolean {
         val navController = this.findNavController(R.id.nav_host_fragment)
         return navController.navigateUp() || super.onSupportNavigateUp()
+    }
+
+    override fun attachBaseContext(newBase: Context?) {
+        super.attachBaseContext(getLocaleAppliedContext(newBase))
     }
 
     override fun onDestroy() {
@@ -48,4 +87,25 @@ class MainActivity : BaseActivity() {
 
         super.onDestroy()
     }
+
+    fun applyWallpaperBackgroundColor(value: Int = PrefManager.systemWallpaperAlpha) {
+        if (PrefManager.systemWallpaper) {
+            val color = (value shl 24) + if (ThemeUtils.isNightMode(this)) {
+                0x00000000
+            } else {
+                0x00FFFFFF
+            }
+
+            window.setBackgroundDrawable(color.toDrawable())
+        }
+    }
+
+    private fun getLocaleAppliedContext(context: Context?): Context? {
+        val config = hmaApp.resources.configuration
+        config.setLocale(ConfigUtils.getLocale())
+
+        return context?.createConfigurationContext(config)
+    }
+
+    private fun isNightModeEnabled(config: Configuration?) = config?.uiMode?.and(Configuration.UI_MODE_NIGHT_MASK)
 }
